websocket-proxy/pyproject.toml
[project]
name = "websocket-proxy"
version = "0.1.0"
description = "Foxglove WebSocket proxy with message transformations"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "numpy>=2.0.0",
    "pointcloud2>=0.2.3",
    "websockets>=15.0.1",
    "typing-extensions>=4.12.2",
    "portable-ffmpeg>=0.1.0",
    "pillow>=12.0.0",
    "mcap-ros2-support-fast",
    "websocket-bridge",
    "rich>=14.2.0",
]

[project.scripts]
bridge = "websocket_proxy.__main__:main"

[build-system]
requires = ["uv_build>=0.8.9,<0.9.0"]
build-backend = "uv_build"

[tool.uv.sources]
mcap-ros2-support-fast = { workspace = true }
websocket-bridge = { workspace = true }

[dependency-groups]
example = [
    "foxglove-sdk>=0.15.3",
]
